# Conflicts:
#	opus/application/apps/ui/templates/base.html
#	opus/application/static_media/js/browse.js
#	opus/application/static_media/js/hash.js
#	opus/application/static_media/js/opus.js

+++ opus/application/apps/cart/views.py
@@ -39,8 +39,7 @@ from dictionary.models import Definitions
 from results.views import (get_search_results_chunk,
                            labels_for_slugs)
 from search.models import ObsGeneral
-from search.views import (get_param_info_by_slug,
-                          url_to_search_params,
+from search.views import (url_to_search_params,
                           get_user_query_table)
 from cart.models import Cart
 from tools.app_utils import *
@@ -69,10 +68,13 @@ def api_view_cart(request):
     This is a PRIVATE API.
 
     Format: __cart/view.html
-
-    For HTML format, returns the left side of the Selections page.
     """
     api_code = enter_api_call('api_view_cart', request)
+
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
 
     session_id = get_session_id(request)
 
@@ -128,6 +130,11 @@ def api_cart_status(request):
     """
     api_code = enter_api_call('api_cart_status', request)
 
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
+
     session_id = get_session_id(request)
 
     reqno = get_reqno(request)
@@ -171,7 +178,17 @@ def api_get_cart_csv(request):
     """
     api_code = enter_api_call('api_get_cart_csv', request)
 
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
+
     column_labels, page = _csv_helper(request, api_code)
+    if column_labels is None:
+        ret = Http404(settings.HTTP404_UNKNOWN_SLUG)
+        exit_api_call(api_code, ret)
+        raise ret
+
     ret = csv_response('data', page, column_labels)
 
     exit_api_call(api_code, ret)
@@ -197,11 +214,16 @@ def api_edit_cart(request, **kwargs):
     """
     api_code = enter_api_call('api_edit_cart', request)
 
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
+
     session_id = get_session_id(request)
 
     try:
         action = kwargs['action']
-    except KeyError:
+    except KeyError: # pragma: no cover
         exit_api_call(api_code, None)
         raise Http404
 
@@ -229,7 +251,7 @@ def api_edit_cart(request, **kwargs):
             err = _edit_cart_range(request, session_id, action, api_code)
         elif action == 'addall':
             err = _edit_cart_addall(request, session_id, api_code)
-        else:
+        else: # pragma: no cover
             assert False
 
     download = request.GET.get('download', 0)
@@ -264,6 +286,11 @@ def api_reset_session(request):
     """
     api_code = enter_api_call('api_reset_session', request)
 
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
+
     session_id = get_session_id(request)
 
     sql = 'DELETE FROM '+connection.ops.quote_name('cart')
@@ -292,6 +319,11 @@ def api_create_download(request, opus_id=None):
                urlonly=1 (optional) means to not zip the actual data products
     """
     api_code = enter_api_call('api_create_download', request)
+
+    if not request or request.GET is None:
+        ret = Http404(settings.HTTP404_NO_REQUEST)
+        exit_api_call(api_code, ret)
+        raise ret
 
     url_file_only = request.GET.get('urlonly', 0)
 
@@ -850,6 +882,10 @@ def _csv_helper(request, api_code=None):
 def _create_csv_file(request, csv_file_name, api_code=None):
     "Create a CSV file containing the cart data."
     column_labels, page = _csv_helper(request, api_code)
+    if column_labels is None:
+        ret = Http404(settings.HTTP404_UNKNOWN_SLUG)
+        exit_api_call(api_code, ret)
+        raise ret
 
     with open(csv_file_name, 'a') as csv_file:
         wr = csv.writer(csv_file)
